Fix Grammar Alternative Set Size (#4)
We are kind of abusing ANTLR here by having such a large alternative group for all the HUGO gene names, and it seems to break down and stop recognizing names towards the end of the list. By breaking down the alternative group into smaller chunks it seems to work fine.